Restructure files and add unit test outline for processing inputs
