+++ .idea/workspace.xml
@@ -2,12 +2,16 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="2438161e-a8e0-4f2b-8fb8-3d04d4ebeb05" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/resources/invalidFileFormat.txt" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/resources/invalidGrid.txt" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/resources/invalidMovementRequest.txt" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/resources/invalidStartingPosition.txt" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/resources/validInputExample.txt" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/main/FileInputProcessor.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/main/Main.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/test/FileInputProcessorTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/MarsRover.iml" beforeDir="false" afterPath="$PROJECT_DIR$/MarsRover.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/com/platform45/marsrover/Grid.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/main/Grid.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/com/platform45/marsrover/Main.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/com/platform45/marsrover/Position.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/main/Position.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/com/platform45/marsrover/Rover.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/com/platform45/marsrover/main/Rover.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/test/com/platform45/marsrover/MainTest.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -18,6 +22,7 @@
     <option name="RECENT_TEMPLATES">
       <list>
         <option value="Class" />
+        <option value="JUnit5 Test Class" />
       </list>
     </option>
   </component>
@@ -85,6 +90,13 @@
     <option name="showLibraryContents" value="true" />
   </component>
   <component name="PropertiesComponent">
+    <property name="Downloaded.Files.Path.Enabled" value="false" />
+    <property name="Repository.Attach.Annotations" value="false" />
+    <property name="Repository.Attach.JavaDocs" value="false" />
+    <property name="Repository.Attach.Sources" value="false" />
+    <property name="com.intellij.testIntegration.createTest.CreateTestDialog.defaultLibrary" value="JUnit5" />
+    <property name="com.intellij.testIntegration.createTest.CreateTestDialog.defaultLibrarySuperClass.JUnit5" value="" />
+    <property name="create.test.in.the.same.root" value="true" />
     <property name="last_opened_file_path" value="$PROJECT_DIR$" />
   </component>
   <component name="RecentsManager">
@@ -92,12 +104,19 @@
       <recent name="C:\Users\ljvan\Documents\Sandra's stuff\Job Option Info\Platform45\MarsRover" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\ljvan\Documents\Sandra's stuff\Job Option Info\Platform45\MarsRover\src\com\platform45\marsrover\test" />
       <recent name="C:\Users\ljvan\Documents\Sandra's stuff\Job Option Info\Platform45\MarsRover\test\com\platform45\marsrover" />
     </key>
+    <key name="CreateTestDialog.Recents.Supers">
+      <recent name="" />
+    </key>
+    <key name="CreateTestDialog.RecentsKey">
+      <recent name="com.platform45.marsrover" />
+    </key>
   </component>
-  <component name="RunManager">
+  <component name="RunManager" selected="JUnit.FileInputProcessorTest.testProcessFileForValidInput">
     <configuration name="Main" type="Application" factoryName="Application" temporary="true">
-      <option name="MAIN_CLASS_NAME" value="com.platform45.marsrover.Main" />
+      <option name="MAIN_CLASS_NAME" value="com.platform45.marsrover.main.Main" />
       <module name="MarsRover" />
       <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
       <RunnerSettings RunnerId="Run" />
@@ -106,6 +125,45 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
+    <configuration name="FileInputProcessorTest.testProcessFileForValidInput" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="MarsRover" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="com.platform45.marsrover.test.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="com.platform45.marsrover.test" />
+      <option name="MAIN_CLASS_NAME" value="com.platform45.marsrover.test.FileInputProcessorTest" />
+      <option name="METHOD_NAME" value="testProcessFileForValidInput" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration name="FileInputProcessorTest.test_JUnit" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="MarsRover" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="com.platform45.marsrover.test.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="com.platform45.marsrover.test" />
+      <option name="MAIN_CLASS_NAME" value="com.platform45.marsrover.test.FileInputProcessorTest" />
+      <option name="METHOD_NAME" value="test_JUnit" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <recent_temporary>
+      <list>
+        <item itemvalue="JUnit.FileInputProcessorTest.testProcessFileForValidInput" />
+        <item itemvalue="JUnit.FileInputProcessorTest.test_JUnit" />
+        <item itemvalue="Application.Main" />
+      </list>
+    </recent_temporary>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
