+++ src/com/platform45/marsrover/main/Rover.java
@@ -1,4 +1,6 @@
-package com.platform45.marsrover;
+package com.platform45.marsrover.main;
+
+import com.platform45.marsrover.main.Position;
 
 /**
  * The rover, which has a position on a grid representing a spot in the physical space of a plateau on Mars.
