--- test/com/platform45/marsrover/MainTest.java
@@ -1,8 +1,0 @@
-package com.platform45.marsrover;
-
-public class Main {
-
-    public static void main(String[] args) {
-	// write your code here
-    }
-}
